Avoid warnings about implicit integer conversions that change the signedness of an integer value (-Wsign-conversion).

+++ Classes/ASIInputStream.m
@@ -144,7 +144,7 @@ static NSLock *readLock = nil;
 	[readLock unlock];
 	NSInteger rv = [stream read:buffer maxLength:toRead];
 	if (rv > 0)
-		[ASIHTTPRequest incrementBandwidthUsedInLastSecond:rv];
+		[ASIHTTPRequest incrementBandwidthUsedInLastSecond:(NSUInteger)rv];
 	return rv;
 }
